pyproject.toml
[tool.poetry]
name = "location-bot"
description = ""
version = "1.0.0"
authors = ["OpenAlcoholics <team@openalcoholics.group>"]
license = "MIT"
packages = [
    { include = "app", from = "src" }
]

[[tool.poetry.source]]
name = "internal"
url = "https://pypi.blindfolded.surgery"
priority = "explicit"

[tool.poetry.dependencies]
bs-config = { version = "1.0.0", source = "internal" }
certifi = "2023.7.22"
httpx = "^0.24.0"
python = "~3.11"
python-telegram-bot = "20.5"
sentry-sdk = "^1.14.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
mypy = "^1.5.0"
pytest = "^7.4.0"
ruff = "0.1.4"
types-requests = "^2.28.11"

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
warn_return_any = false

[[tool.mypy.overrides]]
module = "tests.*"
strict = false
check_untyped_defs = true
disallow_untyped_defs = false

[tool.ruff]
extend-select = [
    "I001",
]

[tool.ruff.isort]
known-first-party = ["az"]
